+++ netket/operator/__init__.py
@@ -18,7 +18,7 @@ from ._discrete_operator import DiscreteOperator
 from ._discrete_operator_jax import DiscreteJaxOperator
 
 from ._pauli_strings import PauliStrings, PauliStringsJax
-from ._local_operator import LocalOperator
+from ._local_operator import LocalOperator, LocalOperatorJax
 from ._graph_operator import GraphOperator
 from ._lazy import Adjoint, Transpose, Squared
 from ._heisenberg import Heisenberg
