+++ netket/operator/_local_operator/base.py
@@ -20,13 +20,11 @@ from textwrap import dedent
 
 import numpy as np
 import jax.numpy as jnp
-import numba
 from scipy.sparse import issparse
 
 from netket.hilbert import AbstractHilbert
 from netket.utils.types import DType, Array
 from netket.utils.numbers import dtype as _dtype, is_scalar
-from netket.errors import concrete_or_error, NumbaOperatorGetConnDuringTracingError
 
 from .._discrete_operator import DiscreteOperator
 from .._lazy import Transpose
@@ -36,7 +34,6 @@ from .helpers import (
     _multiply_operators,
     cast_operator_matrix_dtype,
 )
-from .compile_helpers import pack_internals
 from .convert import local_operators_to_pauli_strings
 
 if TYPE_CHECKING:
@@ -57,7 +54,7 @@ def _is_sorted(a):
     return True
 
 
-class LocalOperator(DiscreteOperator):
+class LocalOperatorBase(DiscreteOperator):
     """A custom local operator. This is a sum of an arbitrary number of operators
     acting locally on a limited set of k quantum numbers (i.e. k-local,
     in the quantum information sense).
@@ -169,8 +166,6 @@ class LocalOperator(DiscreteOperator):
         return self._size
 
     @property
-    # A way to cache the property depending on modifications of self._operators is described here:
-    # https://stackoverflow.com/questions/48262273/python-bookkeeping-dependencies-in-cached-attributes-that-might-change
     def is_hermitian(self) -> bool:
         """Returns true if this operator is hermitian."""
         # TODO: (VolodyaCO) I guess that if we have an operator with diagonal elements equal to 1j*C+Y, some complex constant, and
@@ -219,7 +214,7 @@ class LocalOperator(DiscreteOperator):
         if not np.can_cast(self.dtype, dtype, casting="same_kind"):
             raise ValueError(f"Cannot cast {self.dtype} to {dtype}")
 
-        new = LocalOperator(self.hilbert, constant=self.constant, dtype=dtype)
+        new = type(self)(self.hilbert, constant=self.constant, dtype=dtype)
         new.mel_cutoff = self.mel_cutoff
 
         if dtype == self.dtype:
@@ -264,13 +259,13 @@ class LocalOperator(DiscreteOperator):
     def __neg__(self):
         return -1 * self
 
-    def __add__(self, other: Union["LocalOperator", numbers.Number]):
+    def __add__(self, other: Union["LocalOperatorBase", numbers.Number]):
         op = self.copy(dtype=jnp.promote_types(self.dtype, _dtype(other)))
         op = op.__iadd__(other)
         return op
 
     def __iadd__(self, other):
-        if isinstance(other, LocalOperator):
+        if isinstance(other, LocalOperatorBase):
             if self.hilbert != other.hilbert:
                 return NotImplemented
 
@@ -346,7 +341,7 @@ class LocalOperator(DiscreteOperator):
         return NotImplemented
 
     def __imatmul__(self, other):
-        if not isinstance(other, LocalOperator):
+        if not isinstance(other, LocalOperatorBase):
             return NotImplemented
 
         if not np.can_cast(other.dtype, self.dtype, casting="same_kind"):
@@ -356,14 +351,14 @@ class LocalOperator(DiscreteOperator):
 
         return self._op_imatmul_(other)
 
-    def _op__matmul__(self, other: "LocalOperator") -> "LocalOperator":
-        if not isinstance(other, LocalOperator):
+    def _op__matmul__(self, other: "LocalOperatorBase") -> "LocalOperatorBase":
+        if not isinstance(other, LocalOperatorBase):
             return NotImplemented
         op = self.copy(dtype=jnp.promote_types(self.dtype, _dtype(other)))
         return op._op_imatmul_(other)
 
-    def _op_imatmul_(self, other: "LocalOperator") -> "LocalOperator":
-        if not isinstance(other, LocalOperator):
+    def _op_imatmul_(self, other: "LocalOperatorBase") -> "LocalOperatorBase":
+        if not isinstance(other, LocalOperatorBase):
             return NotImplemented
 
         # (α + ∑ᵢAᵢ)(β + ∑ᵢBᵢ) =
@@ -410,354 +405,11 @@ class LocalOperator(DiscreteOperator):
         self._initialized = False
         self._is_hermitian = None
 
-    def _setup(self, force: bool = False):
-        """Analyze the operator strings and precompute arrays for get_conn inference"""
-        if force or not self._initialized:
-            data = pack_internals(
-                self.hilbert,
-                self._operators_dict,
-                self.constant,
-                self.dtype,
-                self.mel_cutoff,
-            )
-
-            self._acting_on = data["acting_on"]
-            self._acting_size = data["acting_size"]
-            self._diag_mels = data["diag_mels"]
-            self._mels = data["mels"]
-            self._x_prime = data["x_prime"]
-            self._n_conns = data["n_conns"]
-            self._local_states = data["local_states"]
-            self._basis = data["basis"]
-            self._nonzero_diagonal = data["nonzero_diagonal"]
-            self._max_conn_size = data["max_conn_size"]
-            self._initialized = True
-
     @property
     def max_conn_size(self) -> int:
         """The maximum number of non zero ⟨x|O|x'⟩ for every x."""
         self._setup()
         return self._max_conn_size
-
-    def get_conn_flattened(self, x, sections, pad=False):
-        r"""Finds the connected elements of the Operator. Starting
-        from a given quantum number x, it finds all other quantum numbers x' such
-        that the matrix element :math:`O(x,x')` is different from zero. In general there
-        will be several different connected states x' satisfying this
-        condition, and they are denoted here :math:`x'(k)`, for :math:`k=0,1...N_{\mathrm{connected}}`.
-
-        This is a batched version, where x is a matrix of shape (batch_size,hilbert.size).
-
-        Args:
-            x (matrix): A matrix of shape (batch_size,hilbert.size) containing
-                        the batch of quantum numbers x.
-            sections (array): An array of size (batch_size) useful to unflatten
-                        the output of this function.
-                        See numpy.split for the meaning of sections.
-            pad (bool): Whether to use zero-valued matrix elements in order to return all equal sections.
-
-        Returns:
-            matrix: The connected states x', flattened together in a single matrix.
-            array: An array containing the matrix elements :math:`O(x,x')` associated to each x'.
-
-        """
-        self._setup()
-
-        x = concrete_or_error(
-            np.asarray,
-            x,
-            NumbaOperatorGetConnDuringTracingError,
-            self,
-        )
-
-        return self._get_conn_flattened_kernel(
-            x,
-            sections,
-            self._local_states,
-            self._basis,
-            self._constant,
-            self._diag_mels,
-            self._n_conns,
-            self._mels,
-            self._x_prime,
-            self._acting_on,
-            self._acting_size,
-            self._nonzero_diagonal,
-            pad,
-        )
-
-    def _get_conn_flattened_closure(self):
-        self._setup()
-        _local_states = self._local_states
-        _basis = self._basis
-        _constant = self._constant
-        _diag_mels = self._diag_mels
-        _n_conns = self._n_conns
-        _mels = self._mels
-        _x_prime = self._x_prime
-        _acting_on = self._acting_on
-        _acting_size = self._acting_size
-        # workaround my painfully discovered Numba#6979 (cannot use numpy bools in closures)
-        _nonzero_diagonal = bool(self._nonzero_diagonal)
-
-        fun = self._get_conn_flattened_kernel
-
-        def gccf_fun(x, sections):
-            return fun(
-                x,
-                sections,
-                _local_states,
-                _basis,
-                _constant,
-                _diag_mels,
-                _n_conns,
-                _mels,
-                _x_prime,
-                _acting_on,
-                _acting_size,
-                _nonzero_diagonal,
-            )
-
-        return numba.jit(nopython=True)(gccf_fun)
-
-    @staticmethod
-    @numba.jit(nopython=True)
-    def _get_conn_flattened_kernel(
-        x,
-        sections,
-        local_states,
-        basis,
-        constant,
-        diag_mels,
-        n_conns,
-        all_mels,
-        all_x_prime,
-        acting_on,
-        acting_size,
-        nonzero_diagonal,
-        pad=False,
-    ):
-        batch_size = x.shape[0]
-        n_sites = x.shape[1]
-        dtype = all_mels.dtype
-
-        # TODO remove this line when numba 0.53 is dropped 0.54 is minimum version
-        # workaround a bug in Numba arising when NUMBA_BOUNDSCHECK=1
-        constant = constant.item()
-
-        assert sections.shape[0] == batch_size
-
-        n_operators = n_conns.shape[0]
-        xs_n = np.empty((batch_size, n_operators), dtype=np.intp)
-
-        tot_conn = 0
-        max_conn = 0
-
-        for b in range(batch_size):
-            # diagonal element
-            conn_b = 1 if nonzero_diagonal else 0
-
-            # counting the off-diagonal elements
-            for i in range(n_operators):
-                acting_size_i = acting_size[i]
-
-                xs_n[b, i] = 0
-                x_b = x[b]
-                x_i = x_b[acting_on[i, :acting_size_i]]
-                for k in range(acting_size_i):
-                    xs_n[b, i] += (
-                        np.searchsorted(
-                            local_states[i, acting_size_i - k - 1],
-                            x_i[acting_size_i - k - 1],
-                        )
-                        * basis[i, k]
-                    )
-
-                conn_b += n_conns[i, xs_n[b, i]]
-
-            tot_conn += conn_b
-            sections[b] = tot_conn
-
-            if pad:
-                max_conn = max(conn_b, max_conn)
-
-        if pad:
-            tot_conn = batch_size * max_conn
-
-        x_prime = np.empty((tot_conn, n_sites), dtype=x.dtype)
-        mels = np.empty(tot_conn, dtype=dtype)
-
-        c = 0
-        for b in range(batch_size):
-            c_diag = c
-            x_batch = x[b]
-
-            if nonzero_diagonal:
-                mels[c_diag] = constant
-                x_prime[c_diag] = np.copy(x_batch)
-                c += 1
-
-            for i in range(n_operators):
-                if nonzero_diagonal:
-                    mels[c_diag] += diag_mels[i, xs_n[b, i]]
-
-                n_conn_i = n_conns[i, xs_n[b, i]]
-
-                if n_conn_i > 0:
-                    sites = acting_on[i]
-                    acting_size_i = acting_size[i]
-
-                    for cc in range(n_conn_i):
-                        mels[c + cc] = all_mels[i, xs_n[b, i], cc]
-                        x_prime[c + cc] = np.copy(x_batch)
-
-                        for k in range(acting_size_i):
-                            x_prime[c + cc, sites[k]] = all_x_prime[
-                                i, xs_n[b, i], cc, k
-                            ]
-                    c += n_conn_i
-
-            if pad:
-                delta_conn = max_conn - (c - c_diag)
-                mels[c : c + delta_conn].fill(0)
-                x_prime[c : c + delta_conn, :] = np.copy(x_batch)
-                c += delta_conn
-                sections[b] = c
-
-        return x_prime, mels
-
-    def get_conn_filtered(self, x, sections, filters):
-        r"""Finds the connected elements of the Operator using only a subset of operators. Starting
-        from a given quantum number x, it finds all other quantum numbers x' such
-        that the matrix element :math:`O(x,x')` is different from zero. In general there
-        will be several different connected states x' satisfying this
-        condition, and they are denoted here :math:`x'(k)`, for :math:`k=0,1...N_{\mathrm{connected}}`.
-
-        This is a batched version, where x is a matrix of shape (batch_size,hilbert.size).
-
-        Args:
-            x (matrix): A matrix of shape (batch_size,hilbert.size) containing
-                        the batch of quantum numbers x.
-            sections (array): An array of size (batch_size) useful to unflatten
-                        the output of this function.
-                        See numpy.split for the meaning of sections.
-            filters (array): Only operators op(filters[i]) are used to find the connected elements of
-                        x[i].
-
-        Returns:
-            matrix: The connected states x', flattened together in a single matrix.
-            array: An array containing the matrix elements :math:`O(x,x')` associated to each x'.
-
-        """
-        self._setup()
-
-        x = concrete_or_error(
-            np.asarray,
-            x,
-            NumbaOperatorGetConnDuringTracingError,
-            self,
-        )
-
-        return self._get_conn_filtered_kernel(
-            x,
-            sections,
-            self._local_states,
-            self._basis,
-            self._constant,
-            self._diag_mels,
-            self._n_conns,
-            self._mels,
-            self._x_prime,
-            self._acting_on,
-            self._acting_size,
-            filters,
-        )
-
-    @staticmethod
-    @numba.jit(nopython=True)
-    def _get_conn_filtered_kernel(
-        x,
-        sections,
-        local_states,
-        basis,
-        constant,
-        diag_mels,
-        n_conns,
-        all_mels,
-        all_x_prime,
-        acting_on,
-        acting_size,
-        filters,
-    ):
-        batch_size = x.shape[0]
-        n_sites = x.shape[1]
-        dtype = all_mels.dtype
-
-        assert filters.shape[0] == batch_size and sections.shape[0] == batch_size
-
-        # TODO remove this line when numba 0.53 is dropped 0.54 is minimum version
-        # workaround a bug in Numba arising when NUMBA_BOUNDSCHECK=1
-        constant = constant.item()
-
-        n_operators = n_conns.shape[0]
-        xs_n = np.empty((batch_size, n_operators), dtype=np.intp)
-
-        tot_conn = 0
-
-        for b in range(batch_size):
-            # diagonal element
-            tot_conn += 1
-
-            # counting the off-diagonal elements
-            i = filters[b]
-
-            assert i < n_operators and i >= 0
-            acting_size_i = acting_size[i]
-
-            xs_n[b, i] = 0
-            x_b = x[b]
-            x_i = x_b[acting_on[i, :acting_size_i]]
-            for k in range(acting_size_i):
-                xs_n[b, i] += (
-                    np.searchsorted(
-                        local_states[i, acting_size_i - k - 1],
-                        x_i[acting_size_i - k - 1],
-                    )
-                    * basis[i, k]
-                )
-
-            tot_conn += n_conns[i, xs_n[b, i]]
-            sections[b] = tot_conn
-
-        x_prime = np.empty((tot_conn, n_sites))
-        mels = np.empty(tot_conn, dtype=dtype)
-
-        c = 0
-        for b in range(batch_size):
-            c_diag = c
-            mels[c_diag] = constant
-            x_batch = x[b]
-            x_prime[c_diag] = np.copy(x_batch)
-            c += 1
-
-            i = filters[b]
-            # Diagonal part
-            mels[c_diag] += diag_mels[i, xs_n[b, i]]
-            n_conn_i = n_conns[i, xs_n[b, i]]
-
-            if n_conn_i > 0:
-                sites = acting_on[i]
-                acting_size_i = acting_size[i]
-
-                for cc in range(n_conn_i):
-                    mels[c + cc] = all_mels[i, xs_n[b, i], cc]
-                    x_prime[c + cc] = np.copy(x_batch)
-
-                    for k in range(acting_size_i):
-                        x_prime[c + cc, sites[k]] = all_x_prime[i, xs_n[b, i], cc, k]
-                c += n_conn_i
-
-        return x_prime, mels
 
     def __repr__(self):
         ao = self.acting_on
@@ -766,12 +418,3 @@ class LocalOperator(DiscreteOperator):
         if len(acting_str) > 55:
             acting_str = f"#acting_on={len(ao)} locations"
         return f"{type(self).__name__}(dim={self.hilbert.size}, {acting_str}, constant={self.constant}, dtype={self.dtype})"
-
-    def to_jax_operator(self) -> "LocalOperatorJax":  # noqa: F821
-        """
-        Returns the jax-compatible version of this operator, which is an
-        instance of :class:`netket.operator.LocalOperatorJax`.
-        """
-        from .jax import LocalOperatorJax
-
-        return LocalOperatorJax.from_numba_operator(self)
